Install pandoc via debian package
We've found that brew is no longer present by default in ubuntu workers.
This replaces the pandoc installation via brew with a direct
installation from the Debian package included in Pandoc's GitHub release
assets.

An advantage of this approach is that we can now specifically set the
pandoc version being used, and update it based on
https://github.com/jgm/pandoc/releases